requirements.txt
Flask==2.0.1
pytesseract==0.3.8
Pillow==8.2.0
flask-cors==3.0.10
